# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Osmosis Sidecar Query Server Example API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/pools": {
            "get": {
                "description": "Returns a list of pools if the IDs parameter is not given. Otherwise,\nit batch fetches specific pools by the given pool IDs parameter.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get pool(s) information",
                "operationId": "get-pools",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated list of pool IDs to fetch, e.g., '1,2,3'",
                        "name": "IDs",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum pool liquidity cap",
                        "name": "min_liquidity_cap",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pool(s) details",
                        "schema": {
                            "type": "array",
                            "items": {}
                        }
                    }
                }
            }
        },
        "/pools/canonical-orderbook": {
            "get": {
                "description": "Returns the canonical orderbook pool ID for the given base and quote.\nif the pool ID is not found for the given pair, it returns an error.\nif the base or quote denom are not provided, it returns an error.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get canonical orderbook pool ID for the given base and quote.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Base denom",
                        "name": "base",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Quote denom",
                        "name": "quote",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Canonical Orderbook Pool ID for the given base and quote",
                        "schema": {
                            "type": "struct"
                        }
                    }
                }
            }
        },
...
    },
...
}